added /lib/redis/errors.rb file; Redis::ProtocolError inherited from Redis::Error;
