+++ lib/redis.rb
@@ -1,21 +1,6 @@
 require "monitor"
 
 class Redis
-  class ProtocolError < RuntimeError
-    def initialize(reply_type)
-      super(<<-EOS.gsub(/(?:^|\n)\s*/, " "))
-      Got '#{reply_type}' as initial reply byte.
-      If you're running in a multi-threaded environment, make sure you
-      pass the :thread_safe option when initializing the connection.
-      If you're in a forking environment, such as Unicorn, you need to
-      connect to Redis after forking.
-      EOS
-    end
-  end
-
-  class Error < RuntimeError
-  end
-
   module DisableThreadSafety
     def synchronize
       yield
@@ -1195,3 +1180,4 @@ require "redis/client"
 require "redis/pipeline"
 require "redis/subscribe"
 require "redis/compat"
+require "redis/errors"
